refactor: Split the class
